airflow/requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-common-compat==1.13.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.30.4
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.11.0
apache-airflow-task-sdk==1.1.7
argcomplete==3.6.3
asgiref==3.11.1
attrs==25.4.0
babel==2.18.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.6
cadwyn==5.4.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.82.0
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.20
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenback==1.3.0
greenlet==3.3.1
grpcio==1.78.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
methodtools==0.4.7
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
more-itertools==10.8.0
msgspec==0.20.0
natsort==8.4.0
numpy==1.26.4
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.4
pendulum==3.2.0
pillow==12.1.0
pluggy==1.6.0
prettytable==3.17.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
pyparsing==3.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rich-argparse==1.7.2
rich-toolkit==0.18.1
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scikit-surprise==1.1.4
scipy==1.17.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.5.3
websockets==16.0
Werkzeug==3.1.5
wirerope==1.0.0
wrapt==2.1.1
zipp==3.23.0
